+++ FlowCrypt/src/main/java/com/flowcrypt/email/ui/adapter/RecipientsRecyclerViewAdapter.kt
@@ -8,15 +8,14 @@ package com.flowcrypt.email.ui.adapter
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageButton
-import android.widget.ImageView
-import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.flowcrypt.email.R
 import com.flowcrypt.email.database.entity.RecipientEntity
+import com.flowcrypt.email.databinding.ContactItemBinding
 import com.flowcrypt.email.extensions.gone
+import com.flowcrypt.email.extensions.invisible
 import com.flowcrypt.email.extensions.visible
 import com.flowcrypt.email.extensions.visibleOrGone
 
@@ -28,15 +27,13 @@ import com.flowcrypt.email.extensions.visibleOrGone
 class RecipientsRecyclerViewAdapter(
   val isDeleteEnabled: Boolean = true,
   private val onRecipientActionsListener: OnRecipientActionsListener? = null
-) :
-  ListAdapter<RecipientEntity.WithPgpMarker, RecipientsRecyclerViewAdapter.ViewHolder>(
-    DiffUtilCallBack()
-  ) {
-
+) : ListAdapter<RecipientEntity.WithPgpMarker, RecipientsRecyclerViewAdapter.ViewHolder>(
+  DIFF_CALLBACK
+) {
   override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-    val view =
+    return ViewHolder(
       LayoutInflater.from(parent.context).inflate(R.layout.contact_item, parent, false)
-    return ViewHolder(view)
+    )
   }
 
   override fun onBindViewHolder(holder: ViewHolder, position: Int) {
@@ -44,42 +41,38 @@ class RecipientsRecyclerViewAdapter(
   }
 
   inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-    private val tVName: TextView = itemView.findViewById(R.id.tVName)
-    private val tVEmail: TextView = itemView.findViewById(R.id.tVEmail)
-    private val tVOnlyEmail: TextView = itemView.findViewById(R.id.tVOnlyEmail)
-    private val iBtDeleteContact: ImageButton = itemView.findViewById(R.id.iBtDeleteContact)
-    private val imageViewPgp: ImageView = itemView.findViewById(R.id.imageViewPgp)
+    private val binding = ContactItemBinding.bind(itemView)
 
     fun bind(
       recipientEntity: RecipientEntity.WithPgpMarker,
       onRecipientActionsListener: OnRecipientActionsListener?
     ) {
       if (recipientEntity.name.isNullOrEmpty()) {
-        tVName.gone()
-        tVEmail.gone()
-        tVOnlyEmail.visible()
-        tVOnlyEmail.text = recipientEntity.email
-        tVEmail.text = null
-        tVName.text = null
+        binding.tVName.gone()
+        binding.tVEmail.gone()
+        binding.tVOnlyEmail.visible()
+        binding.tVOnlyEmail.text = recipientEntity.email
+        binding.tVEmail.text = null
+        binding.tVName.text = null
       } else {
-        tVName.visible()
-        tVEmail.visible()
-        tVOnlyEmail.gone()
-        tVEmail.text = recipientEntity.email
-        tVName.text = recipientEntity.name
-        tVOnlyEmail.text = null
+        binding.tVName.visible()
+        binding.tVEmail.visible()
+        binding.tVOnlyEmail.gone()
+        binding.tVEmail.text = recipientEntity.email
+        binding.tVName.text = recipientEntity.name
+        binding.tVOnlyEmail.text = null
       }
 
       if (isDeleteEnabled) {
-        iBtDeleteContact.visible()
-        iBtDeleteContact.setOnClickListener {
+        binding.iBtDeleteContact.visible()
+        binding.iBtDeleteContact.setOnClickListener {
           onRecipientActionsListener?.onDeleteRecipient(recipientEntity)
         }
       } else {
-        iBtDeleteContact.gone()
+        binding.iBtDeleteContact.invisible()
       }
 
-      imageViewPgp.visibleOrGone(recipientEntity.hasPgp)
+      binding.imageViewPgp.visibleOrGone(recipientEntity.hasPgp)
 
       itemView.setOnClickListener {
         onRecipientActionsListener?.onRecipientClick(recipientEntity)
@@ -92,17 +85,17 @@ class RecipientsRecyclerViewAdapter(
     fun onRecipientClick(recipientEntityWithPgpMarker: RecipientEntity.WithPgpMarker)
   }
 
-  class DiffUtilCallBack : DiffUtil.ItemCallback<RecipientEntity.WithPgpMarker>() {
-    override fun areItemsTheSame(
-      oldItem: RecipientEntity.WithPgpMarker,
-      newItem: RecipientEntity.WithPgpMarker
-    ) =
-      oldItem.email == newItem.email
+  companion object {
+    private val DIFF_CALLBACK = object : DiffUtil.ItemCallback<RecipientEntity.WithPgpMarker>() {
+      override fun areItemsTheSame(
+        oldItem: RecipientEntity.WithPgpMarker,
+        newItem: RecipientEntity.WithPgpMarker
+      ) = oldItem.email == newItem.email
 
-    override fun areContentsTheSame(
-      oldItem: RecipientEntity.WithPgpMarker,
-      newItem: RecipientEntity.WithPgpMarker
-    ) =
-      oldItem == newItem
+      override fun areContentsTheSame(
+        oldItem: RecipientEntity.WithPgpMarker,
+        newItem: RecipientEntity.WithPgpMarker
+      ) = oldItem == newItem
+    }
   }
 }
